fix(391): use older browser-friendly selector
This walks back the usage of `:is()` in a `document.querySelector` in
order to support older browsers that don't support that CSS selector.

There is no real penalty for doing one over the other; the `:is()`
provided a clearer syntax for writing the selector, and we now
dynamically generate it when the module is first evaluated.