fix(375): NaN invalid CSS width issue
Fixes #375

When an element (or it's "container" concept) has a width and height of
`0`, the `getScale` function tries to do some math that results in
`Infinity`. When it then tried to multiply a dimension of `0` by the resulting
scale (`Infinity`), it returned `NaN`, so we would end up with values
like `{ width: NaN, height: NaN }` that were used in `style` tags on an
`<img />`.

To resolve this, if the values for `containerHeight` or `containerWidth`
are falsy (e.g. `null`, `undefined`, or `0`) for any reason, we bail
from `getScale()` and return a scale of `1`.