fix(406): SVG missing parent error
This fixes an issue where we were unsafely assuming an SVG's parent
element existed, and trying to access a bounding rect property on it
failed, for getting the bounding client rect failed silently. In
essence, `undefined.left` doesn't work.

The fix is to rely on the `imgEl`'s rect if the parent isn't available.