fix(430): ESC intermittently not working in Safari
There's a weird issue, probably related to Safari's `:focus-visible`
behavior or their `<dialog>` implementation, where the `<dialog>`
element's button doesn't receive focus when it's opened. It could also
be a race condition because of the animation.

Regardless, let's try listening for `Escape` at the `document` level
when we zoom an image. We need to make sure we provide the `useCapture`
option and get to the event first so we can intercept it and don't
trigger the default `<dialog>` `Escape` behavior so we can animate
zooming out.