fix: img SVGs with no dimensions
If an `<img />` element has a `src` that is an SVG that has no `width`
nor `height` defined, we got into a weird spot where the loaded element
had `naturalWidth` and `naturalHeight` of `0`. This can differ between
browsers, but most notably in Firefox. In Firefox, since this is `0`,
the object fit calculations were dividing by zero, leading the scaling
factor to go from its correct amount (scale to the browser window) to
not scaling whatsoever and just centering the image on the screen at the
same size it was rendered at. The fix here is to do what we do for other
calculation functions and fallback to the image's rect measurements.