fix: ghost not being styled right
This was a weird one. I don't know how the ghost element is working
correctly in Chrome & Firefox, but in Safari, `[data-rmiz]` having
`position: relative;` makes its ghost child, which is `position:
absolute;` be offset quite a bit, and it's even worse the longer a page
is.

A quick solution is to simply remove that style, but I noticed that
there was some funkyness & inconsistency with how the ghost rendered in
a few situations. Instead of trying to get the ghost's styles _before_
we render (which I knew was unwise, but I did it anyway), this work
moves that to a more after-the-render way of doing things. It also takes
into account pretty much any change that can happen to the `<Zoom>`
children now.